# internal/api/v1/api.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.2 DO NOT EDIT.
package api

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// CreatePurchaseRequest defines model for CreatePurchaseRequest.
type CreatePurchaseRequest struct {
	Data struct {
		Attributes struct {
			Quantity int `json:"quantity"`
		} `json:"attributes"`
		Relationships struct {
			TicketOption TicketOptionRelationship `json:"ticket_option"`
			User         UserRelationship         `json:"user"`
		} `json:"relationships"`
		Type string `json:"type"`
	} `json:"data"`
}

// CreateTicketOptionRequest defines model for CreateTicketOptionRequest.
type CreateTicketOptionRequest struct {
	Data struct {
		Attributes struct {
			Allocation int `json:"allocation"`

			// BucketCount Optional number of capacity buckets for write sharding. Defaults to 1 when omitted. Must be <= allocation and <= 32.
			BucketCount *int    `json:"bucket_count,omitempty"`
			Description *string `json:"description,omitempty"`
			Name        string  `json:"name"`
		} `json:"attributes"`
		Type string `json:"type"`
	} `json:"data"`
}

// ErrorDocument defines model for ErrorDocument.
type ErrorDocument struct {
	Errors []ErrorObject `json:"errors"`
}

// ErrorObject defines model for ErrorObject.
type ErrorObject struct {
	Code   *string `json:"code,omitempty"`
	Detail *string `json:"detail,omitempty"`
	Source *struct {
		Pointer *string `json:"pointer,omitempty"`
	} `json:"source,omitempty"`
	Status string `json:"status"`
	Title  string `json:"title"`
}

// PurchaseDocument defines model for PurchaseDocument.
type PurchaseDocument struct {
	Data PurchaseResource `json:"data"`
}

// PurchaseResource defines model for PurchaseResource.
type PurchaseResource struct {
	Attributes struct {
		Quantity int `json:"quantity"`
	} `json:"attributes"`
	Id            openapi_types.UUID `json:"id"`
	Relationships struct {
		TicketOption TicketOptionRelationship `json:"ticket_option"`
		User         UserRelationship         `json:"user"`
	} `json:"relationships"`
	Type string `json:"type"`
}

// TicketOptionDocument defines model for TicketOptionDocument.
type TicketOptionDocument struct {
	Data TicketOptionResource `json:"data"`
}

// TicketOptionRelationship defines model for TicketOptionRelationship.
type TicketOptionRelationship struct {
	Data struct {
		Id   openapi_types.UUID `json:"id"`
		Type string             `json:"type"`
	} `json:"data"`
}

// TicketOptionResource defines model for TicketOptionResource.
type TicketOptionResource struct {
	Attributes struct {
		Allocation  int     `json:"allocation"`
		Description *string `json:"description,omitempty"`
		Name        string  `json:"name"`
	} `json:"attributes"`
	Id   openapi_types.UUID `json:"id"`
	Type string             `json:"type"`
}

// UserRelationship defines model for UserRelationship.
type UserRelationship struct {
	Data struct {
		Id   openapi_types.UUID `json:"id"`
		Type string             `json:"type"`
	} `json:"data"`
}

// CreatePurchaseApplicationVndAPIPlusJSONRequestBody defines body for CreatePurchase for application/vnd.api+json ContentType.
type CreatePurchaseApplicationVndAPIPlusJSONRequestBody = CreatePurchaseRequest

// CreateTicketOptionApplicationVndAPIPlusJSONRequestBody defines body for CreateTicketOption for application/vnd.api+json ContentType.
type CreateTicketOptionApplicationVndAPIPlusJSONRequestBody = CreateTicketOptionRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create Purchase
	// (POST /v1/purchases)
	CreatePurchase(w http.ResponseWriter, r *http.Request)
	// Create ticket option
	// (POST /v1/ticket_options)
	CreateTicketOption(w http.ResponseWriter, r *http.Request)
	// Get ticket option
	// (GET /v1/ticket_options/{ticketOptionID})
	GetTicketOption(w http.ResponseWriter, r *http.Request, ticketOptionID openapi_types.UUID)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// CreatePurchase operation middleware
func (siw *ServerInterfaceWrapper) CreatePurchase(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreatePurchase(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateTicketOption operation middleware
func (siw *ServerInterfaceWrapper) CreateTicketOption(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateTicketOption(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTicketOption operation middleware
func (siw *ServerInterfaceWrapper) GetTicketOption(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "ticketOptionID" -------------
	var ticketOptionID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "ticketOptionID", r.PathValue("ticketOptionID"), &ticketOptionID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "ticketOptionID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTicketOption(w, r, ticketOptionID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/v1/purchases", wrapper.CreatePurchase)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/v1/ticket_options", wrapper.CreateTicketOption)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/v1/ticket_options/{ticketOptionID}", wrapper.GetTicketOption)

	return m
}

type CreatePurchaseRequestObject struct {
	Body *CreatePurchaseApplicationVndAPIPlusJSONRequestBody
}

type CreatePurchaseResponseObject interface {
	VisitCreatePurchaseResponse(w http.ResponseWriter) error
}

type CreatePurchase201ApplicationVndAPIPlusJSONResponse PurchaseDocument

func (response CreatePurchase201ApplicationVndAPIPlusJSONResponse) VisitCreatePurchaseResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/vnd.api+json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type CreatePurchase400ApplicationVndAPIPlusJSONResponse ErrorDocument

func (response CreatePurchase400ApplicationVndAPIPlusJSONResponse) VisitCreatePurchaseResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/vnd.api+json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type CreatePurchase404ApplicationVndAPIPlusJSONResponse ErrorDocument

func (response CreatePurchase404ApplicationVndAPIPlusJSONResponse) VisitCreatePurchaseResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/vnd.api+json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type CreateTicketOptionRequestObject struct {
	Body *CreateTicketOptionApplicationVndAPIPlusJSONRequestBody
}

type CreateTicketOptionResponseObject interface {
	VisitCreateTicketOptionResponse(w http.ResponseWriter) error
}

type CreateTicketOption201ApplicationVndAPIPlusJSONResponse TicketOptionDocument

func (response CreateTicketOption201ApplicationVndAPIPlusJSONResponse) VisitCreateTicketOptionResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/vnd.api+json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type CreateTicketOption400ApplicationVndAPIPlusJSONResponse ErrorDocument

func (response CreateTicketOption400ApplicationVndAPIPlusJSONResponse) VisitCreateTicketOptionResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/vnd.api+json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetTicketOptionRequestObject struct {
	TicketOptionID openapi_types.UUID `json:"ticketOptionID"`
}

type GetTicketOptionResponseObject interface {
	VisitGetTicketOptionResponse(w http.ResponseWriter) error
}

type GetTicketOption200ApplicationVndAPIPlusJSONResponse TicketOptionDocument

func (response GetTicketOption200ApplicationVndAPIPlusJSONResponse) VisitGetTicketOptionResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/vnd.api+json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetTicketOption404ApplicationVndAPIPlusJSONResponse ErrorDocument

func (response GetTicketOption404ApplicationVndAPIPlusJSONResponse) VisitGetTicketOptionResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/vnd.api+json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Create Purchase
	// (POST /v1/purchases)
	CreatePurchase(ctx context.Context, request CreatePurchaseRequestObject) (CreatePurchaseResponseObject, error)
	// Create ticket option
	// (POST /v1/ticket_options)
	CreateTicketOption(ctx context.Context, request CreateTicketOptionRequestObject) (CreateTicketOptionResponseObject, error)
	// Get ticket option
	// (GET /v1/ticket_options/{ticketOptionID})
	GetTicketOption(ctx context.Context, request GetTicketOptionRequestObject) (GetTicketOptionResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// CreatePurchase operation middleware
func (sh *strictHandler) CreatePurchase(w http.ResponseWriter, r *http.Request) {
	var request CreatePurchaseRequestObject

	var body CreatePurchaseApplicationVndAPIPlusJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreatePurchase(ctx, request.(CreatePurchaseRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreatePurchase")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreatePurchaseResponseObject); ok {
		if err := validResponse.VisitCreatePurchaseResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreateTicketOption operation middleware
func (sh *strictHandler) CreateTicketOption(w http.ResponseWriter, r *http.Request) {
	var request CreateTicketOptionRequestObject

	var body CreateTicketOptionApplicationVndAPIPlusJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateTicketOption(ctx, request.(CreateTicketOptionRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateTicketOption")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateTicketOptionResponseObject); ok {
		if err := validResponse.VisitCreateTicketOptionResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetTicketOption operation middleware
func (sh *strictHandler) GetTicketOption(w http.ResponseWriter, r *http.Request, ticketOptionID openapi_types.UUID) {
	var request GetTicketOptionRequestObject

	request.TicketOptionID = ticketOptionID

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetTicketOption(ctx, request.(GetTicketOptionRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetTicketOption")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetTicketOptionResponseObject); ok {
		if err := validResponse.VisitGetTicketOptionResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}